+++ src/DiscoPoP/runner/DiscoPoPCMakeWorkflowRunner.ts
@@ -1,102 +1,11 @@
 import * as fs from 'fs'
-import * as vscode from 'vscode'
+import { DiscoPoPRunner } from './DiscoPoPRunner'
 import { exec } from 'child_process'
-import { DiscoPoPResults } from './classes/DiscoPoPResults'
-import { Config } from '../Utils/Config'
-import { EventEmitter } from 'events'
-import { UIPrompts } from '../Utils/UIPrompts'
-import { DiscoPoPParser } from './DiscoPoPParser'
+import { CancelToken } from '../../Utils/CancelToken'
+import { Config } from '../../Utils/Config'
+import { DiscoPoPResults } from '../classes/DiscoPoPResults'
 
-export type CancelToken = {
-    emitter: EventEmitter // emits "cancel" event when cancellation is requested
-    cancellationRequested: boolean
-}
-
-/** provides functions to run the DiscoPoP tools */
-export class DPRunner {
-    public constructor(public readonly dotDiscoPoP: string) {}
-
-    public get discopopVersion(): string {
-        return '' // TODO
-    }
-
-    public get discopopBuild(): string {
-        return Config.discopopBuild() // TODO replace with discopop_config_provider
-    }
-
-    public async runExplorer(cancelToken: CancelToken): Promise<void> {
-        let outerResolve: () => void
-        return new Promise<void>((resolve, reject) => {
-            outerResolve = resolve
-            const childProcess = exec(
-                `discopop_explorer`,
-                {
-                    cwd: this.dotDiscoPoP,
-                },
-                (err, stdout, stderr) => {
-                    if (err) {
-                        reject(
-                            new Error(
-                                'discopop_explorer failed: ' + err.message
-                            )
-                        )
-                    } else {
-                        resolve()
-                    }
-                }
-            )
-            cancelToken.emitter.on('cancel', async () => {
-                console.log(
-                    'DiscoPoPRunner::cancellation requested::discopop_explorer'
-                )
-                childProcess.kill() // SIGINT or SIGTERM?
-                outerResolve?.()
-            })
-        })
-    }
-    public async runPatchGenerator(cancelToken: CancelToken): Promise<void> {
-        let outerResolve: () => void
-        return new Promise<void>((resolve, reject) => {
-            outerResolve = resolve
-            const childProcess = exec(
-                `discopop_patch_generator`,
-                {
-                    cwd: this.dotDiscoPoP,
-                },
-                (err, stdout, stderr) => {
-                    if (err) {
-                        reject(
-                            new Error(
-                                'discopop_patch_generator failed: ' +
-                                    err.message +
-                                    '\n' +
-                                    stderr
-                            )
-                        )
-                    } else {
-                        resolve()
-                    }
-                }
-            )
-            cancelToken.emitter.on('cancel', async () => {
-                console.log(
-                    'DiscoPoPRunner::cancellation requested::discopop_patch_generator'
-                )
-                childProcess.kill() // SIGINT or SIGTERM?
-                outerResolve?.()
-            })
-        })
-    }
-
-    public async parse(): Promise<DiscoPoPResults> {
-        return DiscoPoPParser.parse(this.dotDiscoPoP)
-    }
-
-    public async runPatchApplicator(): Promise<void> {}
-    public async runOptimizer(): Promise<void> {}
-}
-
-export class DPRunnerCMake {
+export class DiscoPoPCMakeWorkflowRunner {
     /**
      *
      * @param projectDirectory
@@ -135,7 +44,7 @@ export class DPRunnerCMake {
         requestConfirmation: (message: string) => Promise<boolean>,
         cancelToken: CancelToken
     ): Promise<DiscoPoPResults> {
-        const dpRunner = new DPRunner(this.dotDiscoPoP)
+        const dpRunner = new DiscoPoPRunner(this.dotDiscoPoP)
 
         reportMessage('Preparing...', 0)
         await this.prepare(requestConfirmation)
@@ -301,95 +210,5 @@ export class DPRunnerCMake {
                 outerResolve?.()
             })
         })
-    }
-}
-
-export class DPRunnerUI {
-    public constructor(
-        public readonly projectDirectory: string,
-        public readonly executableName: string,
-        public readonly executableArguments: string = '',
-        public readonly buildDirectory?: string,
-        public readonly dotDiscoPoP?: string
-    ) {
-        if (!this.buildDirectory) {
-            this.buildDirectory = projectDirectory + '/build/DiscoPoP'
-        }
-
-        if (!this.dotDiscoPoP) {
-            this.dotDiscoPoP = projectDirectory + 'build/.discopop'
-        }
-    }
-
-    public async run() {
-        const dpRunnerCMake = new DPRunnerCMake(
-            this.projectDirectory,
-            this.executableName,
-            this.executableArguments,
-            this.buildDirectory,
-            this.dotDiscoPoP
-        )
-
-        await vscode.window.withProgress(
-            {
-                location: vscode.ProgressLocation.Notification,
-                title: 'Running DiscoPoP',
-                cancellable: true,
-            },
-            async (progress, token) => {
-                const cancelTokenWrapper: CancelToken = {
-                    emitter: new EventEmitter(),
-                    cancellationRequested: false,
-                }
-                cancelTokenWrapper.emitter.on('cancel', () => {
-                    cancelTokenWrapper.cancellationRequested = true
-                })
-                const cancellationDisposable = token.onCancellationRequested(
-                    () => {
-                        cancelTokenWrapper.emitter.emit('cancel')
-                    }
-                )
-
-                const reportMessageWrapper = (
-                    message: string,
-                    nesting: number
-                ) => {
-                    // all progress reports are logged to the console
-                    console.log(
-                        `DiscoPoPRunner: ${'-'.repeat(nesting)} ${message}`
-                    )
-
-                    // only top-level progress reports are shown in the UI
-                    if (nesting === 0) {
-                        progress.report({
-                            message: message,
-                        })
-                    }
-                }
-
-                const reportProgressWrapper = (progressValue: number) => {
-                    progress.report({
-                        increment: progressValue,
-                    })
-                }
-
-                const requestConfirmationWrapper = async (message: string) => {
-                    return UIPrompts.actionConfirmed(message)
-                }
-
-                try {
-                    await dpRunnerCMake.run(
-                        reportMessageWrapper,
-                        reportProgressWrapper,
-                        requestConfirmationWrapper,
-                        cancelTokenWrapper
-                    )
-                } catch (e: any) {
-                    vscode.window.showErrorMessage(e.message || e)
-                } finally {
-                    cancellationDisposable.dispose()
-                }
-            }
-        )
     }
 }
